增加主键递增的插入方法

+++ src/com/neusoft/pm/dept/dao/DeptDao.java
@@ -11,7 +11,7 @@ import java.util.List;
 public interface DeptDao {
     List<Dept> queryList(Dept dept);
 
-    int insert(Dept dept);
+    Long insert(Dept dept);
 
     int update(Dept dept);
